[64] Other/Checksum: b969d6f2bd69b73b753e7d7bce1e4669

+++ gitp.py
@@ -17,12 +17,17 @@ try:
 		return("\033[92m" + string + "\033[0m")
 
 	refs = Popen("git remote", shell=True, stdout=PIPE).stdout.read().rstrip()
+	removed_list = list()
 
 	for ref in refs.splitlines():
 		print(green("*") + " Pulling '" + ref + "'")
 		raw_data = Popen("git pull --quiet " + ref + " master", shell=True, stdout=PIPE).stdout.read().rstrip()
+
 		for data in raw_data.splitlines():
 			print(red("*") + " " + data)
+
+			if data.split()[0] == "Removing":
+				removed_list.append(data.split()[0])
 
 	if not access("list", R_OK):
 		print(blue("*") + " Creating list")
@@ -37,20 +42,29 @@ try:
 
 	print(blue("*") + " Removing non-existing files")
 	for lists in open("list", "r"):
+		removed = False
 		filename = lists.rstrip()[:-33]
 		list_checksum = lists.rstrip().split()[-32:]
+
 		if not access(filename, R_OK):
-			raw_data = Popen("git rm --quiet --cached --force " + filename.replace(" ", "\ "), shell=True, stdout=PIPE).stdout.read().rstrip()
-			for data in raw_data.splitlines():
-				print(red("*") + " " + data)
+			for removed_file in removed_list:
+				if removed_file == filename:
+					removed = True
+
+			if not removed:
+				raw_data = Popen("git rm --quiet --cached --force " + filename.replace(" ", "\ "), shell=True, stdout=PIPE).stdout.read().rstrip()
+
+				for data in raw_data.splitlines():
+					print(red("*") + " " + data)
 
 	print(blue("*") + " Clearing list")
 	file_writer = open("list", "w")
 	file_writer.write("")
-	
 	print(blue("*") + " Adding existing files")
+
 	for root, dirs, files in sorted(walk(".")):
 		files.sort()
+
 		for name in files:
 			if not path.join(root[2:], name) == "list" and not path.join(root[2:], name) == "version" and not path.join(root[2:], name) == ".gitignore" and not path.join(root[2:], name).startswith(".git/"):
 				ignored = False
@@ -62,6 +76,7 @@ try:
 
 				if not ignored:
 					file_checksum = md5("").hexdigest()
+
 					if access(path.join(root[2:], name), R_OK):
 						file_reader = open(path.join(root[2:], name), "r")
 						file_content = file_reader.read()
@@ -71,19 +86,18 @@ try:
 					file_writer.write(path.join(root[2:], name) + " " + file_checksum + "\n")
 
 	file_writer.close()
-
 	print(blue("*") + " Creating listhash")
 	listfile = open("list", "r")
 	listhash = md5(listfile.read()).hexdigest()
 	listfile.close()
-
 	print(blue("*") + " Comparing listhashes")
+
 	if listprehash != listhash:
 		revision = Popen("git log --oneline | wc -l", shell=True, stdout=PIPE).stdout.read().rstrip()
 		print(blue("*") + " Getting new revision")
 		revision = str(int(revision) + 1)
-
 		print(blue("*") + " Creating version")
+
 		if len(revision) > 3:
 			version = revision[:-3] + "." + revision[-3:]
 		elif len(revision) == 3:
@@ -97,19 +111,20 @@ try:
 		f = open("version", "w")
 		f.write(version)
 		f.close()
-
 		raw_data = Popen("git add .", shell=True, stdout=PIPE).stdout.read().rstrip()
+
 		for data in raw_data.splitlines():
 			print(red("*") + " " + data)
 
 		repository = "origin master"
-
 		hascommit = False
+
 		if len(argv) > 1:
 			if len(argv) > 3:
 				if argv[-3] == "-r":
 					repository = ' '.join(argv[-2:])
 					commit = ' '.join(argv[1:-3])
+
 					if len(argv) > 4:
 						hascommit = True
 				else:
@@ -128,11 +143,13 @@ try:
 
 		print(blue("*") + " Committing '" + commit + "'")
 		raw_data = Popen("git commit --quiet --message '[" + revision + "] " + commit + "' --signoff", shell=True, stdout=PIPE).stdout.read().rstrip()
+
 		for data in raw_data.splitlines():
 			print(red("*") + " " + data)
 
 		print(green("*") + " Pushing to '" + repository + "'")
 		raw_data = Popen("git push --quiet " + repository, shell=True, stdout=PIPE).stdout.read().rstrip()
+
 		for data in raw_data.splitlines():
 			print(red("*") + " " + data)
 	else:
